feat: refactor memory persistence
